requirements.txt
Flask==0.12.2
paho-mqtt==1.3.0